Type safety for iOS

+++ include/qstabilizerhybrid.hpp
@@ -155,7 +155,7 @@ protected:
     void CheckShots(unsigned shots, const bitCapInt& m, real1_f partProb, const std::vector<bitCapInt>& qPowers,
         std::vector<real1_f>& rng, F fn)
     {
-        for (int64_t shot = rng.size() - 1U; shot >= 0; --shot) {
+        for (unsigned int shot = rng.size() - 1U; shot >= 0; --shot) {
             if (rng[shot] >= partProb) {
                 break;
             }
